refactor: remove sample resource e muda a pasta do jaxrs
Remoção do sample resource api e alteração do package do JAXRS config
