--- src/main/java/themovies/SampleResource.java
@@ -1,22 +1,0 @@
-package themovies;
-
-import javax.inject.Inject;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.core.Response;
-
-import org.eclipse.microprofile.config.inject.ConfigProperty;
-
-@Path("sample")
-public class SampleResource {
-
-    @Inject
-    @ConfigProperty(name = "message")
-    private String message;
-
-    @GET
-    public Response message() {
-        return Response.ok(message).build();
-    }
-
-}
